moved handler & util
